# back/public/api-docs.yaml
openapi: 3.0.0
info:
  title: 'API ProjetPro'
  description: "API REST pour l'application ProjetPro avec système d'authentification sécurisé.\n\n## Fonctionnalités principales:\n- Authentification avec protection anti-force brute\n- Gestion des tokens CSRF\n- Upload d'images vers Azure Blob Storage\n- Logging automatique des tentatives de connexion\n- Système de blocage par IP et par login\n\n## Sécurité:\n- Protection CSRF obligatoire\n- Tokens d'authentification JWT\n- Blocage automatique après échecs répétés\n\n## Headers requis:\n- `X-CSRF-Token`: Token CSRF pour les requêtes POST\n- `Authorization: Bearer <token>`: Token JWT pour l'authentification\n- `Content-Type: application/json`\n"
  contact:
    name: 'Support API'
    email: support@projetpro.com
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
    description: 'Serveur de développement'
  -
    url: /api
    description: 'API Production'
paths:
...
  /api/machines/:
    get:
      tags:
        - Machine
      summary: 'Liste des machines'
      description: 'Retourne la liste de toutes les machines'
      operationId: get_app_machines_all_machines
      responses:
        '200':
          description: 'Liste des machines'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MachineResponseDTO'
    post:
      tags:
        - Machine
      summary: 'Créer une machine'
      operationId: post_app_machines_create_machine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MachineCreateRequestDTO'
      responses:
        '201':
          description: 'Machine créée'
        '400':
          description: 'Erreur de validation'
  '/api/machines/{materielId}':
    get:
      tags:
        - Machine
      summary: 'Récupérer une machine'
      description: "Retourne les informations d'une machine par son ID"
      operationId: get_app_machines_machine_by_id
      parameters:
        -
          name: materielId
          in: path
          description: 'ID de la machine'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Données de la machine'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MachineResponseDTO'
    put:
      tags:
        - Machine
      summary: 'Mettre à jour une machine'
      description: 'Met à jour une machine existante'
      operationId: put_app_machines_update_machine
      parameters:
        -
          name: materielId
          in: path
          description: 'ID de la machine'
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MachineUpdateRequestDTO'
      responses:
        '200':
          description: 'Machine mise à jour'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MachineResponseDTO'
        '400':
          description: 'Erreur de validation'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean }
                  message: { type: string }
                type: object
        '404':
          description: 'Erreur lors de la mise à jour'
          content:
            application/json:
              schema:
                properties:
                  error: { type: string }
                type: object
...
components:
  schemas:
...
    MachineResponseDTO:
      description: Machine
      required:
        - id
        - name
        - visible
        - image
        - description
        - forum
        - infoMachines
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'PC Portable'
        visible:
          type: boolean
          example: true
        image:
          type: string
          example: image.png
        description:
          type: string
          example: 'Ordinateur performant'
        forum:
          properties:
            id:
              type: integer
              example: 1
          type: object
        infoMachines:
          type: array
          items:
            properties:
              id:
                type: integer
                example: 1
              text:
                type: string
                example: 'RAM 16Go'
              type:
                type: string
                example: hardware
            type: object
      type: object
    MachineCreateRequestDTO:
      description: 'Machine creation request'
      required:
        - nom
        - image
        - infoMachines
      properties:
        nom:
          type: string
          example: 'PC Portable'
        image:
          type: string
          example: image.png
        infoMachines:
          type: array
          items:
            $ref: App\DTO\InfoMachineBaseDTO
      type: object
    MachineUpdateRequestDTO:
      description: 'Machine update request'
      required:
        - nom
        - image
        - description
        - infoMachines
      properties:
        nom:
          type: string
          example: 'PC Portable'
        image:
          type: string
          example: image.png
        description:
          type: string
          example: 'Ordinateur performant'
        infoMachines:
          type: array
          items:
            $ref: App\DTO\InfoMachineUpdateDTO
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      description: "Token JWT d'authentification"
      bearerFormat: JWT
      scheme: bearer
    csrfToken:
      type: apiKey
      description: 'Token CSRF pour la protection contre les attaques CSRF'
      name: X-CSRF-Token
      in: header
security:
  -
    bearerAuth: []
  -
    csrfToken: []
...